split pauli string helper methods and the DomainWallVariable class into their own utility python files
